Initial commits for porting to ROCm.
Currently, the transform_engine/common module can compile with F8 header
files integrated. bf16 is disabled for now due to gaps between rocm and
cuda. gemm part is commented out for now. And the inline asm will also
need to be replaced.

There are still issues to be addressed for transformer/pytorch module.